Remove second comand palette provider file
